+++ src/component/project/main/mainComponent/Branding/component.tsx
@@ -1,0 +1,63 @@
+import styled from "@emotion/styled";
+
+export const Text1 = styled.p`
+  position: absolute;
+  left: 310px;
+  top: 449px;
+  font-family: Inter;
+  font-size: 30px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+  text-align: left;
+  color: #808080;
+`;
+
+export const Text2 = styled.p`
+   position: absolute;
+   left :310px ;
+   top :415px ;
+   font-family :Inter ;
+   font-size :100px ;
+   font-style :normal ;
+   font-weight :500 ;
+   line-height :normal ;
+   text-align :left ;
+   color:#808080; 
+`;
+
+export const CustomButton = styled.button`
+    width: 382px;
+    height: 68px;
+    position: absolute;
+    left: 310px;
+    top: 634px;
+    border: 1px solid #a5a5a5;
+    border-radius: 87px;
+    font-size: 25px;
+    font-weight: bold;
+    text-align: center;
+    color: #808080;
+    margin-top: 28px; 
+`;
+
+export const RoundedDiv = styled.div`
+     position: absolute;
+     right: 594px;
+     top: 282px;
+     width: 334px;
+     height: 334px;
+     background-color: #D9D9D9;
+     border-radius: 50%; 
+`;
+
+export const RightBottomDiv = styled.div`
+       position: absolute;
+       right: 0;
+       bottom: 0;
+       width: 821px;
+       height: 821px;
+       background-color: #D9D9D9;  
+       border-radius: 50%;   
+`;
+
