+++ src/component/project/main/mainComponent/Consider/consider.tsx
@@ -1,6 +1,4 @@
 import React from "react";
-import styled from "styled-components";
-
 
 export const Consider = () => {
   return (
