+++ src/component/project/main/mainComponent/Howwecanhelp/component.tsx
@@ -1,0 +1,20 @@
+import styled from "@emotion/styled";
+
+export const Container = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center; /* Updated to center vertically */
+  height: 100vh;
+`;
+
+export const ConsiderBoxItem = styled.div`
+  width: 951px;
+  flex-shrink: 0;
+  border-radius: 60px;
+  background: #D9D9D9;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin: 0 10px;
+`;
